Put back auto in upload file
